chore: use regional api
